requirements.txt
streamlit==1.38.0
langchain==0.2.16
faiss-cpu==1.8.0
pymupdf==1.24.10
sentence-transformers==3.1.1
torch==2.4.1